15-jwt-auth/requirements.txt
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
